+++ src/helix/bits/get_bits_leaderboard.rs
@@ -154,6 +154,7 @@ impl RequestGet for GetBitsLeaderboardRequest {
             },
             pagination: None,
             request,
+            total: Some(response.total),
         })
     }
 }