+++ src/helix/channels/modify_channel_information.rs
@@ -131,6 +131,7 @@ impl RequestPatch for ModifyChannelInformationRequest {
             },
             pagination: None,
             request,
+            total: None,
         })
     }
 }