+++ src/helix/eventsub/create_eventsub_subscription.rs
@@ -173,6 +173,8 @@ impl<E: EventSubscription> helix::RequestPost for CreateEventSubSubscriptionRequ
             },
             pagination: None,
             request,
+            // helix::Response total is generally the total number of results, not what the total for this endpoint means. Thus, we set it to None.
+            total: None,
         })
     }
 }